feat(tts): pass properties to the pipeline engine at startup from a file
This is how we will be able to test TTS engines that require configuration on startup - put a file called 'tts-engine-properties.json' in the app data directory. If the file is present, the Pipeline UI will pass those properties to the Pipeline engine, e.g. org.daisy.pipeline.tts.azure.key or other credentials.

This is a temporary solution until there is an API for setting properties in the engine.