fix(#198): resolve naming conflict with options vs stylesheet parameters
Introduce new job request property: "stylesheetParameterOptions" to store stylesheet parameter values.

Visual control for stylesheet parameters is still based on the "optional" array, which is probably badly named, but can represent an optional input, an actual option, or a stylesheet parameter. But the structure where the value is stored is "stylesheetParameterOptions", which helps avoid naming conflicts with script options.