fix(#188): settings dialog is not always shown
The settings dialog is like any other window now, not always on top.

But now we need a way to bring it forward when it is hidden behind other windows. E.g. user selects file->settings and the settings window is already open but is behind some other windows, we need to show it.