feat(#293): avoid large font getting too large
The problem with enlarging the font is that the headings and tabs scale up too big, and it crowds out all the other text on the screen. This adjustment limits how big the large font can get, allowing the smaller font to scale up better.

Also the various font sizes used throughout are streamlined in this commit.