fix(#329): add browse files and browse folders buttons to new job screen
This lets windows users browse for folders/files separately (as required by the platform/electron).

Mac users will see a single button for browsing for folders/files.

And because potentially two buttons are there, we can't have the drag area be dual-purpose (drag + click to browse), so it is now just for drag/drop. To browse with the file selector, users should use the buttons. Will post screenshots later.